Builds sequences of parameters for testing mate pairing equations

Builds sequences of parameters for testing mate pairing equations

error_check_build_Ptest_df(include.W.0 = FALSE, include.K.0 = FALSE,
  seqW = seq(1, 100, 30), seqK = seq(1, 250, 50), seqP.cgg = NA,
  seqP.cgp = NA, seqP.cpg = NA, seqP.cpp = NA, seqP.kgg = NA,
  seqP.kgp = NA, seqP.kpg = NA, seqP.kpp = NA)

Arguments

include.W.0
xxx

include.K.0
xxx

seqW
xxx

seqK
xxx

seqP.cgg
xxx

seqP.cgp
xxx

seqP.cpg
xxx

seqP.cpp
xxx

seqP.kgg
xxx

seqP.kgp
xxx

seqP.kpg
xxx

seqP.kpp
xxx

Examples

P.test.df <- error_check_build_Ptest_df()
dim(P.test.df)

#> [1] 6400   14