Make paramater ranges

Generates a dataframe defining the minimum maximum values of each parameter; these are then turned into a matrix of all combinations by the function makeParamCombos.df()

makeParamRanges.df(scenario. = NA, figure = NA, gamma. = c(5, 5),
  c. = c(1, 1), K.bc. = c(1000, 1000), K.bk. = c(10000, 10000),
  K.wg. = c(900, 900), S.w.mg. = c(0.8, 0.8), S.w.mp. = c(0.8, 0.8),
  S.w.fg. = c(0.8, 0.8), S.w.fp. = c(0.8, 0.8), S.m.mg. = c(0.9, 0.9),
  S.m.mp. = c(0.8, 0.8), S.m.fg. = c(0.9, 0.9), S.m.fp. = c(0.8, 0.8),
  R.base.rate. = c(1.8, 1.8), R.hab.effect. = c(0.5, 0.5), co. = c(1, 1),
  f. = c(0.5, 0.5), S.b.mc. = c(0.95, 0.95), S.b.mk. = c(0.85, 0.85),
  S.b.md. = c(0.8, 0.8), S.b.fc. = c(0.95, 0.95), S.b.fk. = c(0.85, 0.85),
  S.f.mc. = c(0.8, 0.8), S.f.mk. = c(0.75, 0.75), S.f.md. = c(0.8, 0.8),
  S.f.fc. = c(0.8, 0.8), S.f.fk. = c(0.75, 0.75), S.y.mc. = c(0.8, 0.8),
  S.y.mk. = c(0.75, 0.75), S.y.fc. = c(0.8, 0.8), S.y.fk. = c(0.75, 0.75))

Arguments

figure
generate figure. Currently only figures 28.3 and 28.4 are implemented

gamma.
xx

c.
xxx

K.bc.
xxx

K.bk.
xxx

K.wg.
xxx

S.w.mg.
xxx

S.w.mp.
xxx

S.w.fg.
xxx

S.w.fp.
xxx

S.m.mg.
do they call this s.smg?

S.m.mp.
xxx

S.m.fg.
xxx

S.m.fp.
xxx

R.base.rate.
xxx

R.hab.effect.
effects of pair breeding in poor habitat

co.
magnitude of carry over effect; c 1 equals no carry over

f.
sex ratio

S.b.mc.
males in sour.c.e habitat

S.b.mk.
males in sin.k. habitat

S.b.md.
"drain" males

S.b.fc.
xxx

S.b.fk.
xxx

S.f.mc.
males from sour.c.e

S.f.mk.
xxx

S.f.md.
drain males have higher surv b/c they don't have costs of repro

S.f.fc.
xxx

S.f.fk.
xxx

S.y.mc.
xxx

S.y.mk.
xxx

S.y.fc.
xxx

S.y.fk.
xxx

scenario
Difference scenarios defined by Runge & Marra

Value

Dataframe with 2 columns, minimum and maxium values for each parameter

param.ranges 30 x 2 dataframe of the min. and max. parameter values that will be considered. Often values are fixed.

References

Runge, MC and PP Marra. 2004. Modeling seasonal interactions in the population dynamics of migratory birds. In Greenberg, R and PP Marra, eds. Birds of two worlds. Johns Hopkins University Press, Baltimore.

Examples

param.ranges <- makeParamRanges.df(scenario. = "winter.limited")
head(param.ranges)

#>            min     max
#> gamma      5.0     5.0
#> co         1.0     1.0
#> K.bc     800.0   800.0
#> K.bk   10000.0 10000.0
#> K.wg     485.0   485.0
#> S.w.mg     0.8     0.8

param.ranges <- makeParamRanges.df(figure = 28.3)
param.ranges[c("K.bc","K.wg"), ]

#>      min  max
#> K.bc   1 1000
#> K.wg   1 1000