Make combinations of parameters for parameter matrix

This takes a df of parameter ranges generated by paramRanges.df() and applied makeParamSeq() to turn any that vary into a sequence spanning the range. These sequences (if any) are then turned into a matrix of all possible values if none vary, then a single row of data is returned

param_grid(param.seqs = param_seqs(param_ranges(figure = 28.3)))

Arguments

param.seqs
sequence of parameters to use

Value

param.grid Dataframe of parameter combinations

Details

Formerly aliased as makeParamCombos.df

Examples

# "Grid" with no variation = single row of data
param_grid()

#> The dimension of the fully expanded dataframe is: 
#> 100 by 30

#>    gamma co. K.bc  K.bk K.wg S.w.mg S.w.mp S.w.fg S.w.fp S.m.mg S.m.mp S.m.fg
#> 1      5   1    1 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 2      5   1  112 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 3      5   1  223 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 4      5   1  334 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 5      5   1  445 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 6      5   1  556 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 7      5   1  667 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 8      5   1  778 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 9      5   1  889 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 10     5   1 1000 10000    1    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 11     5   1    1 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 12     5   1  112 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 13     5   1  223 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 14     5   1  334 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 15     5   1  445 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 16     5   1  556 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 17     5   1  667 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 18     5   1  778 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 19     5   1  889 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 20     5   1 1000 10000  112    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 21     5   1    1 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 22     5   1  112 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 23     5   1  223 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 24     5   1  334 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 25     5   1  445 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 26     5   1  556 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 27     5   1  667 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 28     5   1  778 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 29     5   1  889 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 30     5   1 1000 10000  223    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 31     5   1    1 10000  334    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 32     5   1  112 10000  334    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#> 33     5   1  223 10000  334    0.8    0.8    0.8    0.8    0.9    0.8    0.9
#>    S.m.fp R.base.rate R.hab.effect   f S.b.mc S.b.mk S.b.md S.b.fc S.b.fk
#> 1     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 2     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 3     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 4     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 5     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 6     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 7     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 8     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 9     0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 10    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 11    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 12    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 13    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 14    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 15    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 16    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 17    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 18    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 19    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 20    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 21    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 22    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 23    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 24    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 25    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 26    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 27    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 28    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 29    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 30    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 31    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 32    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#> 33    0.8         1.8          0.5 0.5   0.95   0.85    0.8   0.95   0.85
#>    S.f.mc S.f.mk S.f.md S.f.fc S.f.fk S.y.mc S.y.mk S.y.fc S.y.fk
#> 1     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 2     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 3     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 4     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 5     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 6     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 7     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 8     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 9     0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 10    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 11    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 12    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 13    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 14    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 15    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 16    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 17    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 18    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 19    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 20    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 21    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 22    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 23    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 24    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 25    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 26    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 27    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 28    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 29    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 30    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 31    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 32    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#> 33    0.8   0.75    0.8    0.8   0.75    0.8   0.75    0.8   0.75
#>  [ reached 'max' / getOption("max.print") -- omitted 67 rows ]

# Grid with variation

##Generate parameters for Figure 28.2
param.ranges <- param_ranges(figure = 28.3)

##generate parameter sequences
param.seqs <- param_seqs(param.ranges = param.ranges)
param.seqs[1:3]

#> $gamma
#> [1] 5
#> 
#> $co.
#> [1] 1
#> 
#> $K.bc
#>  [1]    1  112  223  334  445  556  667  778  889 1000
#> 

##generate grid
param.grid <- param_grid(param.seqs)

#> The dimension of the fully expanded dataframe is: 
#> 100 by 30

## look at snippet of raw data
param.grid[1:10,1:5]

#>    gamma co. K.bc  K.bk K.wg
#> 1      5   1    1 10000    1
#> 2      5   1  112 10000    1
#> 3      5   1  223 10000    1
#> 4      5   1  334 10000    1
#> 5      5   1  445 10000    1
#> 6      5   1  556 10000    1
#> 7      5   1  667 10000    1
#> 8      5   1  778 10000    1
#> 9      5   1  889 10000    1
#> 10     5   1 1000 10000    1

## look at summary of columns that vary
summary(param.grid[,c("K.bc","K.bk","K.wg" )])

#>       K.bc             K.bk            K.wg       
#>  Min.   :   1.0   Min.   :10000   Min.   :   1.0  
#>  1st Qu.: 223.0   1st Qu.:10000   1st Qu.: 223.0  
#>  Median : 500.5   Median :10000   Median : 500.5  
#>  Mean   : 500.5   Mean   :10000   Mean   : 500.5  
#>  3rd Qu.: 778.0   3rd Qu.:10000   3rd Qu.: 778.0  
#>  Max.   :1000.0   Max.   :10000   Max.   :1000.0  

## conver colums that vary to factors then look at them
param.grid$K.bc <- factor(param.grid$K.bc)
param.grid$K.bk <- factor(param.grid$K.bk)
param.grid$K.wg <- factor(param.grid$K.wg)

summary(param.grid[,c("K.bc","K.bk","K.wg" )],10)

#>    K.bc       K.bk       K.wg   
#>  1   :10   10000:100   1   :10  
#>  112 :10               112 :10  
#>  223 :10               223 :10  
#>  334 :10               334 :10  
#>  445 :10               445 :10  
#>  556 :10               556 :10  
#>  667 :10               667 :10  
#>  778 :10               778 :10  
#>  889 :10               889 :10  
#>  1000:10               1000:10