Make paramater ranges

Generates a dataframe defining the minimum maximum values of each parameter; these are then turned into a matrix of all combinations by the function makeParamCombos.df()

param_set(scenario = NA, figure = NA, gamma. = 5, c. = 1,
  K.bc. = 1000, K.bk. = 10000, K.wg. = 900, S.w.mg. = 0.8,
  S.w.mp. = 0.8, S.w.fg. = 0.8, S.w.fp. = 0.8, S.m.mg. = 0.9,
  S.m.mp. = 0.9, S.m.fg. = 0.9, S.m.fp. = 0.9, R.base.rate. = 1.8,
  R.hab.effect. = 0.5, co. = 1, f. = 0.5, S.b.mc. = 0.95,
  S.b.mk. = 0.85, S.b.md. = 0.8, S.b.fc. = 0.95, S.b.fk. = 0.85,
  S.f.mc. = 0.8, S.f.mk. = 0.75, S.f.md. = 0.8, S.f.fc. = 0.8,
  S.f.fk. = 0.75, S.y.mc. = 0.8, S.y.mk. = 0.75, S.y.fc. = 0.8,
  S.y.fk. = 0.75)

Arguments

scenario
Difference scenarios defined by Runge & Marra: "winter" limited, "intermediate", and "spring" limited. If set overides other varibles.

figure
deprecated

gamma.
xx

c.
xxx

K.bc.
xxx

K.bk.
xxx

K.wg.
xxx

S.w.mg.
xxx

S.w.mp.
xxx

S.w.fg.
xxx

S.w.fp.
xxx

S.m.mg.
do they call this s.smg?

S.m.mp.
xxx

S.m.fg.
xxx

S.m.fp.
xxx

R.base.rate.
xxx

R.hab.effect.
effects of pair breeding in poor habitat

co.
magnitude of carry over effect; c 1 equals no carry over

f.
sex ratio

S.b.mc.
males in sour.c.e habitat

S.b.mk.
males in sin.k. habitat

S.b.md.
"drain" males

S.b.fc.
xxx

S.b.fk.
xxx

S.f.mc.
males from sour.c.e

S.f.mk.
xxx

S.f.md.
drain males have higher surv b/c they don't have costs of repro

S.f.fc.
xxx

S.f.fk.
xxx

S.y.mc.
xxx

S.y.mk.
xxx

S.y.fc.
xxx

S.y.fk.
xxx

Value

param.ranges 30 x 2 dataframe of the min. and max. parameter values that will be considered. Often values are fixed and do not vary.

References

Runge, MC and PP Marra. 2004. Modeling seasonal interactions in the population dynamics of migratory birds. In Greenberg, R and PP Marra, eds. Birds of two worlds: The ecology & evolution of migration. Johns Hopkins University Press, Baltimore.

Examples

# Default values
## Print out default values (only 1st 10 show.  Remove "[1:10]" to see all)
param_set()[1:10]

#>   gamma co K.bc  K.bk K.wg S.w.mg S.w.mp S.w.fg S.w.fp S.m.mg
#> 1     5  1 1000 10000  900    0.8    0.8    0.8    0.8    0.9

## Print defaults in a column by taking the transpose using t()
## (only 1st 10 show.  Remove "[1:10]" to see all)
t(param_set()[1:10])

#>         [,1]
#> gamma  5e+00
#> co     1e+00
#> K.bc   1e+03
#> K.bk   1e+04
#> K.wg   9e+02
#> S.w.mg 8e-01
#> S.w.mp 8e-01
#> S.w.fg 8e-01
#> S.w.fp 8e-01
#> S.m.mg 9e-01

# Set parameters for the different scenarios plotted in figure Figure 28.5 in Runge & Marra (2004)
## (note that to actually replicate Fig 28.5 param_rangs() and param_seqs() need to be used)
## to set up full parameter grid with the male dominanc parameter gamma varying.

## winter limited
winter <- param_set(scenario = "winter.limited")

## intermediate scenario
intermediate <- param_set(scenario = "intermediate")

## summer limited scenario
summer <- param_set(scenario = "summer.limited")

## compare the scenarios
### K.bc varies
### k.bk fixed
### K.wg varies
rbind(winter = winter, intermed = intermediate, summer = summer)[, 1:10]

#>          gamma co K.bc  K.bk K.wg S.w.mg S.w.mp S.w.fg S.w.fp S.m.mg
#> winter       5  1  800 10000  485    0.8    0.8    0.8    0.8    0.9
#> intermed     5  1  224 10000  580    0.8    0.8    0.8    0.8    0.9
#> summer       5  1  205 10000  900    0.8    0.8    0.8    0.8    0.9