Run FAC model to equilibirum

Iterates a single parameterization of the Runge & Marra 2004 FAC model until equilbirium is reached. Typically the desired output is the equilibrium population size(s) and all intermediate output is discarded.

runFAC(reps = 350, Ninit = c(10, 0, 10, 0),
  paramsWorking = makeParCombos(), scenario = NA, vary = NA,
  verbose = FALSE, check.eq = TRUE, check.eq.after.i = 20, tol.1 = 3,
  tol.2 = 3, diagnostic.plot = F, ...)

Arguments

reps
how many iterations of model to run before stopping

Ninit
vector of initial abundances at the begining of winter for W.mg, W.mp, W.fg and W.fp

paramsWorking
initila parameters for this run of the model.

scenario
character string representing name of scenario explored in original Runge and Marra paper (not currently implemented 7/10/2018)

vary
...

verbose
verbose output

check.eq
should equilibrium be assessed?

tol.1
tolerance value for diagnostics checks for whether the run of the model has reached equilibrium. Lower values are more stringent; minimum is 1, which makes the model run the full number of iterations

tol.2
secondary tolerance value for checking equilbrium; LARGER values increase stringency

diagnostic.plot
Return a diagnostic plot show population size over time

minimum
number of iterations to run if equilibrium is being checked

Value

A dataframe containing the status of the popualtion at each time step for all parameters

References

Runge, MC & PP Marra 2004. Modeling seasonal interactions in the population dynamics of migratory birds. Birds of two words.