FAC Step 0: Set up survival matrices & other preliminaries

FAC Step 0: Set up survival matrices & other preliminaries

step0_set_up(
  param.set = param_set(),
  iterations = 150,
  use.IBM = F,
  use.IBM.S.b = F
)

Arguments

param.set
Initial parameters and population state for a single run of model to equilibirum

iterations
Number of iterations to run model to allow it to reach equilibrium

use.IBM
Use individual-based modeling implementation of breeding territory aquisition and mate pairing

use.IBM.S.b
Used individual-based modeling implementation of breeding survibal.

Value

runFAC.i list ...

Examples

# Initialize model
runFAC.i <- step0_set_up()

# Look at structure of list
## Elements for model output, matrices, parameters, etc
str(runFAC.i,1)

#> List of 8
#>  $ FAC.out.RM        :'data.frame':	150 obs. of  42 variables:
#>  $ FAC.out.IB        :'data.frame':	150 obs. of  42 variables:
#>  $ FAC.eq.state.RM   : logi NA
#>  $ FAC.eq.state.IB   : logi NA
#>  $ W.mg.diagnostic.df:'data.frame':	150 obs. of  5 variables:
#>  $ param.matrices    :List of 8
#>  $ params.initial    :'data.frame':	1 obs. of  30 variables:
#>  $ use.IBM           : logi FALSE

# Look at top of dataframe to hold output
head(runFAC.i$FAC.out.RM)[,1:5]

#>   W.mg W.mp W.fg W.fp B.mc
#> 1   NA   NA   NA   NA   NA
#> 2   NA   NA   NA   NA   NA
#> 3   NA   NA   NA   NA   NA
#> 4   NA   NA   NA   NA   NA
#> 5   NA   NA   NA   NA   NA
#> 6   NA   NA   NA   NA   NA

# Look at 2nd matrix used in model, winter survival
runFAC.i$param.matrices[2]

#> $S.w
#>      [,1] [,2] [,3] [,4]
#> [1,]  0.8  0.0  0.0  0.0
#> [2,]  0.0  0.8  0.0  0.0
#> [3,]  0.0  0.0  0.8  0.0
#> [4,]  0.0  0.0  0.0  0.8
#> 

runFAC.i$param.matrices$S.w

#>      [,1] [,2] [,3] [,4]
#> [1,]  0.8  0.0  0.0  0.0
#> [2,]  0.0  0.8  0.0  0.0
#> [3,]  0.0  0.0  0.8  0.0
#> [4,]  0.0  0.0  0.0  0.8