Run FAC model to equilibirum

Iterates a single parameterization of the Runge & Marra 2004 FAC model until equilbirium is reached. Typically the desired output is the equilibrium population size(s) and all intermediate output is discarded.

runFAC(iterations = 350, Ninit = c(10, 0, 10, 0),
  param.set = param_set(), scenario = NA, vary = NA,
  verbose = FALSE, check.errors = TRUE, check.errors.in = c("B0.RM",
  "P.cgg", "P.cgp", "P.cpg", "P.cpp", "P.kgg", "P.kgp", "P.kpg", "P.kpp",
  "Y1"), check.eq = TRUE, minimum.i = 20, eq.tol = 6,
  diagnostic.plot = F, return.output = T, save.ts = T, use.IBM = F,
  use.IBM.S.b = F, ...)

Arguments

iterations
how many iterations of model to run before stopping

Ninit
vector of initial abundances at the begining of winter for W.mg, W.mp, W.fg and W.fp

param.set
initila parameters for this run of the model.

scenario
character string representing name of scenario explored in original Runge and Marra paper (not currently implemented 7/10/2018)

vary
...

verbose
verbose general output

check.errors
Run error checks on model subcomponents

check.errors.in
run error checks and report

check.eq
should equilibrium be assessed?

minimum.i
number of iterations to run if equilibrium is being checked

eq.tol
Tolerance for equilibirum check; the number of digistic lambda and variace of lambda should be rounded to when comparing against 1. Larger numbers reduce the likelihood of model passing the equilbirium check

diagnostic.plot
Return a diagnostic plot show population size over time

return.output
Output the full model dataframe?

save.ts
Save the full time series of the model run? If FALSE then only final time point will be returned. Additionally, equilibrium monitoring will not be done and the time series cannot be plotted

use.IBM
use individual based model for habitat and mate acquisition

use.IBM.S.b
use IBM for summer adult survival

...
other arguements passed

Value

A dataframe containing the status of the popualtion at each time step for all parameters

References

Runge, MC and PP Marra. 2004. Modeling seasonal interactions in the population dynamics of migratory birds. In Greenberg, R and PP Marra, eds. Birds of two worlds. Johns Hopkins University Press, Baltimore.

Examples

# Run an FAC
test.FAC <- runFAC(verbose = TRUE)

#> 
#> Model at equilibrium after 196 iterations

# Look at structure of output
str(test.FAC,1)

#> List of 8
#>  $ FAC.out.RM        :'data.frame':	350 obs. of  48 variables:
#>  $ FAC.out.IB        :'data.frame':	350 obs. of  42 variables:
#>  $ FAC.eq.state.RM   :'data.frame':	1 obs. of  48 variables:
#>  $ FAC.eq.state.IB   : logi NA
#>  $ W.mg.diagnostic.df:'data.frame':	350 obs. of  5 variables:
#>  $ param.matrices    :List of 8
#>  $ params.initial    :'data.frame':	1 obs. of  30 variables:
#>  $ use.IBM           : logi FALSE